+++ importlib_metadata/tests/test_api.py
@@ -1,12 +1,16 @@
 import re
 import unittest
+import importlib_metadata
 
 try:
     from collections.abc import Iterator
 except ImportError:
     from collections import Iterator  # noqa: F401
 
-import importlib_metadata
+try:
+    from builtins import str as text
+except ImportError:
+    from __builtin__ import unicode as text
 
 
 class APITests(unittest.TestCase):
@@ -14,13 +18,13 @@ class APITests(unittest.TestCase):
 
     def test_retrieves_version_of_self(self):
         version = importlib_metadata.version('importlib_metadata')
-        assert isinstance(version, str)
+        assert isinstance(version, text)
         assert re.match(self.version_pattern, version)
 
     def test_retrieves_version_of_pip(self):
         # Assume pip is installed and retrieve the version of pip.
         version = importlib_metadata.version('pip')
-        assert isinstance(version, str)
+        assert isinstance(version, text)
         assert re.match(self.version_pattern, version)
 
     def test_for_name_does_not_exist(self):