+++ importlib_metadata/tests/test_main.py
@@ -1,25 +1,21 @@
+# coding: utf-8
+# flake8: noqa (https://gitlab.com/python-devs/importlib_metadata/issues/31)
+
 from __future__ import unicode_literals
 
 import re
-import sys
-import shutil
-import tempfile
+import textwrap
 import unittest
 import importlib
-import contextlib
 import importlib_metadata
 
-try:
-    from contextlib import ExitStack
-except ImportError:
-    from contextlib2 import ExitStack
-
-try:
-    import pathlib
-except ImportError:
-    import pathlib2 as pathlib
-
 from importlib_metadata import _hooks
+from . import fixtures
+
+try:
+    from builtins import str as text
+except ImportError:
+    from __builtin__ import unicode as text
 
 
 class BasicTests(unittest.TestCase):
@@ -28,7 +24,7 @@ class BasicTests(unittest.TestCase):
     def test_retrieves_version_of_pip(self):
         # Assume pip is installed and retrieve the version of pip.
         dist = importlib_metadata.Distribution.from_name('pip')
-        assert isinstance(dist.version, str)
+        assert isinstance(dist.version, text)
         assert re.match(self.version_pattern, dist.version)
 
     def test_for_name_does_not_exist(self):
@@ -60,7 +56,7 @@ class ImportTests(unittest.TestCase):
         self.assertRaises(ValueError, importlib_metadata.resolve, 'bogus.ep')
 
 
-class NameNormalizationTests(unittest.TestCase):
+class NameNormalizationTests(fixtures.SiteDir, unittest.TestCase):
     @staticmethod
     def pkg_with_dashes(site_dir):
         """
@@ -73,22 +69,6 @@ class NameNormalizationTests(unittest.TestCase):
         with metadata.open('w') as strm:
             strm.write('Version: 1.0\n')
         return 'my-pkg'
-
-    @staticmethod
-    @contextlib.contextmanager
-    def site_dir():
-        tmpdir = tempfile.mkdtemp()
-        sys.path[:0] = [tmpdir]
-        try:
-            yield pathlib.Path(tmpdir)
-        finally:
-            sys.path.remove(tmpdir)
-            shutil.rmtree(tmpdir)
-
-    def setUp(self):
-        self.fixtures = ExitStack()
-        self.addCleanup(self.fixtures.close)
-        self.site_dir = self.fixtures.enter_context(self.site_dir())
 
     def test_dashes_in_dist_name_found_as_underscores(self):
         """
@@ -119,3 +99,45 @@ class NameNormalizationTests(unittest.TestCase):
         assert importlib_metadata.version(pkg_name) == '1.0'
         assert importlib_metadata.version(pkg_name.lower()) == '1.0'
         assert importlib_metadata.version(pkg_name.upper()) == '1.0'
+
+
+class NonASCIITests(fixtures.SiteDir, unittest.TestCase):
+    @staticmethod
+    def pkg_with_non_ascii_description(site_dir):
+        """
+        Create minimal metadata for a package with non-ASCII in
+        the description.
+        """
+        metadata_dir = site_dir / 'portend.dist-info'
+        metadata_dir.mkdir()
+        metadata = metadata_dir / 'METADATA'
+        with metadata.open('w', encoding='utf-8') as strm:
+            strm.write('Description: pôrˈtend\n')
+        return 'portend'
+
+    @staticmethod
+    def pkg_with_non_ascii_description_egg_info(site_dir):
+        """
+        Create minimal metadata for an egg-info package with
+        non-ASCII in the description.
+        """
+        metadata_dir = site_dir / 'portend.dist-info'
+        metadata_dir.mkdir()
+        metadata = metadata_dir / 'METADATA'
+        with metadata.open('w', encoding='utf-8') as strm:
+            strm.write(textwrap.dedent("""
+                Name: portend
+
+                pôrˈtend
+                """).lstrip())
+        return 'portend'
+
+    def test_metadata_loads(self):
+        pkg_name = self.pkg_with_non_ascii_description(self.site_dir)
+        meta = importlib_metadata.metadata(pkg_name)
+        assert meta['Description'] == 'pôrˈtend'
+
+    def test_metadata_loads_egg_info(self):
+        pkg_name = self.pkg_with_non_ascii_description_egg_info(self.site_dir)
+        meta = importlib_metadata.metadata(pkg_name)
+        assert meta.get_payload() == 'pôrˈtend\n'